feat: 게임 결과 반환 페이지네이션 서비스

+++ UKIDS/src/main/java/com/modernfamily/ukids/domain/game/quizResult/dto/QuizResultPaginationResponseDto.java
@@ -18,14 +18,14 @@ public class QuizResultPaginationResponseDto {
 
     int totalPage;
 
-    int totalItems;
+    long totalItems;
 
     List<QuizResultResponseDto> quizResults;
 
     UserDto user;
 
     @Builder
-    private QuizResultPaginationResponseDto(int size, int currentPage, int totalPage, int totalItems,
+    private QuizResultPaginationResponseDto(int size, int currentPage, int totalPage, long totalItems,
                                             List<QuizResultResponseDto> quizResults, UserDto user) {
 
         this.size = size;
@@ -36,7 +36,7 @@ public class QuizResultPaginationResponseDto {
         this.user = user;
     }
 
-    public static QuizResultPaginationResponseDto createResponseDto(int size, int currentPage, int totalPage, int totalItems,
+    public static QuizResultPaginationResponseDto createResponseDto(int size, int currentPage, int totalPage, long totalItems,
                                                                     List<QuizResultResponseDto> quizResults, UserDto user) {
         return QuizResultPaginationResponseDto.builder()
                 .size(size)
